Backgrounds?
Added a stary background and a second one to attempt to install a "parallax" like environment. Added a space script with a public variable of parrallax that you can use to moidfy the speed of the stars to simulate space speed. We could use that for gems or to provide variety in levels.